+++ Lib/fontbakery/specifications/googlefonts.py
@@ -67,7 +67,7 @@ METADATA_CHECKS = [
 ]
 
 DESCRIPTION_CHECKS = [
-        'com.google.fonts/check/003' # Does DESCRIPTION file contain broken links?
+        'com.google.fonts/check/description/broken_links' # Does DESCRIPTION file contain broken links?
       , 'com.google.fonts/check/004' # Is this a propper HTML snippet?
       , 'com.google.fonts/check/005' # DESCRIPTION.en_us.html must have more than 200 bytes.
       , 'com.google.fonts/check/006' # DESCRIPTION.en_us.html must have less than 1000 bytes.
@@ -374,10 +374,10 @@ def description(descfile):
 
 
 @check(
-  id = 'com.google.fonts/check/003',
+  id = 'com.google.fonts/check/description/broken_links',
   conditions = ['description']
 )
-def com_google_fonts_check_003(description):
+def com_google_fonts_check_description_broken_links(description):
   """Does DESCRIPTION file contain broken links?"""
   from lxml.html import HTMLParser
   import defusedxml.lxml